REST_APIs/requirements.txt
requests==2.21.0
Flask==1.0.2
flask_restx
marshmallow
pymongo
jsonify
mysql-connector-python